refactor: separate i18nNavigation and i18nRouting to avoid import error with Vitest
