+++ src/components/LocaleSwitcher.tsx
@@ -1,7 +1,8 @@
 'use client';
 
 import type { ChangeEventHandler } from 'react';
-import { routing, usePathname } from '@/libs/i18nNavigation';
+import { usePathname } from '@/libs/i18nNavigation';
+import { routing } from '@/libs/i18nRouting';
 import { useLocale } from 'next-intl';
 import { useRouter } from 'next/navigation';
