+++ src/libs/i18nRouting.ts
@@ -1,0 +1,8 @@
+import { AppConfig } from '@/utils/AppConfig';
+import { defineRouting } from 'next-intl/routing';
+
+export const routing = defineRouting({
+  locales: AppConfig.locales,
+  localePrefix: AppConfig.localePrefix,
+  defaultLocale: AppConfig.defaultLocale,
+});
